Care needed if cookie or crumb contains a slash, easiest to run again....

+++ create_cookie.sh
@@ -1,4 +1,9 @@
-rm history
+#!/bin/bash
+
+if  [ -f history ]; then
+	rm history
+fi
+
 wget --save-cookies cookie_file https://uk.finance.yahoo.com/quote/ANTO.L/history
 
 cookie=$(awk '/yahoo/{print $6"="$7}' cookie_file)
@@ -8,8 +13,11 @@ crumb=$(awk -F, '/CrumbStore/{for(i=1;i<NF;++i){print $i}}' history|sed -n "/Cru
 
 
 line=$(echo string crumb=\"$crumb\",cookie=\"$cookie\"\;)
+
+echo Run again if crumb or cookie has a \\ in it!!!!!!!!!!!!!!
 echo $line
 
 sed -i "/string crumb=/c \$$line" InternetPrice.cs
 sed -i "s/^\\$/\t\t\t/" InternetPrice.cs
 
+rm history
